requirements.txt
numpy
pandas
kagglehub
scikit-learn
matplotlib
evidently
seaborn
ipykernel
dvc
mlflow
PyYAML
ensure==1.0.2
python-box==6.0.2
from_root
dill
typing
dagshub
gradio
nltk
tensorflow
keras
pymongo
spacy==3.6.1
Pillow

-e .